move AddItem and RetrieveItems under trips package
