+++ internal/pkg/trips/add_item.go
