+++ internal/pkg/trips/add_item_test.go
